requirements.txt
asttokens==2.2.1
backcall==0.2.0
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
executing==1.2.0
filelock==3.12.0
Flask==2.2.3
Flask-SocketIO==5.3.4
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.4
hyperframe==5.2.0
idna==2.10
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==0.5.6
msgpack-python==0.5.6
mysql-connector==2.2.9
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pypng==0.20220715.0
python-barcode==0.14.0
python-dateutil==2.8.2
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzbar==0.1.9
pyzmq==25.0.2
qrcode==7.4.2
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
sympy==1.11.1
texttable==1.6.7
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.22.0
wcwidth==0.2.6
Werkzeug==2.2.3
